adapt to unity 2021.2.10f1
Compute some values for price manually to get a somewhat desirable difficulty slope.

Some methods were added and left as public but should ultimately be removed as they were added to output the price and income to the outside world for easy calculations.
Also, the price per tier list should not be initialized inside the tierify function, this was a last minute test.